+++ apps/spruce/src/pages/adminSettings/tabs/GeneralTab/FeatureFlagsTab/FeatureFlagsTab.tsx
@@ -1,6 +1,6 @@
 import { H2 } from "@leafygreen-ui/typography";
-import { AdminSettingsTabRoutes } from "constants/routes";
-import { BaseTab } from "../BaseTab";
+import { AdminSettingsGeneralSection } from "constants/routes";
+import { BaseTab } from "../../BaseTab";
 import { getFormSchema } from "./formSchema";
 import { TabProps } from "./types";
 
@@ -13,7 +13,7 @@ export const FeatureFlagsTab: React.FC<TabProps> = ({ featureFlagsData }) => {
       <BaseTab
         formSchema={formSchema}
         initialFormState={initialFormState}
-        tab={AdminSettingsTabRoutes.FeatureFlags}
+        tab={AdminSettingsGeneralSection.FeatureFlags}
       />
     </>
   );
